refactor into src layout and add optional deps to toml
