+++ pyproject.toml
@@ -11,7 +11,10 @@ authors = [
 readme = "README.md"
 license = {text = "BSD 3-Clause License"}
 dynamic = ["version"]
-dependencies = ["numpy", "cvxpy", "scikit-learn"]
+dependencies = [
+    "numpy >=1.23", "cvxpy >=1.2", "scikit-learn >=1.2",
+    "scipy >=1.9", "joblib", "cvxopt"
+    ]
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Programming Language :: Python :: 3 :: Only",
@@ -25,6 +28,12 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Mathematics",
     "Topic :: Software Development :: Libraries :: Python Modules"
 ]
+
+[project.optional-dependencies]
+docs = ["sphinx>=5.3", "furo", "m2r2"]
+tests = ["pytest >=7.2.0", "pytest-cov >=4.0.0", "coverage"]
+dev = ["pre-commit", "black", "isort", "flake8", "pylint", "pydocstyle", "flake8-pyproject"]
+optional = ["gurobipy"]
 
 # pyproject.toml
 [tool.setuptools_scm]
