+++ src/sparselm/model/_base.py
