+++ src/sparselm/model/_ols.py
