+++ src/sparselm/model/miqp/__init__.py
@@ -1,11 +1,11 @@
 """MIQP based regression estimators."""
 
 
-from sparselm.model.miqp._best_subset import (
+from src.sparselm.model.miqp._best_subset import (
     BestSubsetSelection,
     RidgedBestSubsetSelection,
 )
-from sparselm.model.miqp._regularized_l0 import L1L0, L2L0, RegularizedL0
+from src.sparselm import L1L0, L2L0, RegularizedL0
 
 __all__ = [
     "BestSubsetSelection",
