+++ src/sparselm/model/miqp/_regularized_l0.py
@@ -20,7 +20,7 @@ import cvxpy as cp
 import numpy as np
 from cvxpy.atoms.affine.wraps import psd_wrap
 
-from sparselm.model._base import CVXEstimator
+from src.sparselm.model._base import CVXEstimator
 
 
 # RegularizedL0 now supports orbit-level hierarchy, but now requires groups as a
